Split gfold into gfold and libgfold
Split gfold into gfold and libgfold to increase the number of
compilation units and allow for external use of the library. Keep
libgfold unpublished until logging, error types, etc. is handled
properly.
