+++ bin/gfold/src/display/color.rs
@@ -1,11 +1,11 @@
 //! This module provides a harness for non-trivial displays of information to `stdout`.
 
+use libgfold::Status;
 use std::io;
 use std::io::Write;
 use termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
 
 use crate::config::ColorMode;
-use crate::status::Status;
 
 /// This harness provides methods to write to `stdout`. It maps the internal [`ColorMode`] type to
 /// our dependency's [`ColorChoice`] type due to discrepancies in behavior and naming.
