+++ bin/xtask/src/task/scan.rs
